Remove duplicate tests. Move enum tests into Enum.spec.ts
