+++ src/files/BrsFile.spec.ts
@@ -17,7 +17,7 @@ import { DiagnosticMessages } from '../DiagnosticMessages';
 import type { StandardizedFileEntry } from 'roku-deploy';
 import util, { standardizePath as s } from '../util';
 import PluginInterface from '../PluginInterface';
-import { expectDiagnostics, expectHasDiagnostics, expectZeroDiagnostics, getTestTranspile, trim, expectCompletionsIncludes, expectCompletionsExcludes } from '../testHelpers.spec';
+import { expectDiagnostics, expectHasDiagnostics, expectZeroDiagnostics, getTestTranspile, trim } from '../testHelpers.spec';
 import { ParseMode } from '../parser/Parser';
 import { Logger } from '../Logger';
 
@@ -211,241 +211,6 @@ describe('BrsFile', () => {
                 result = program.getCompletions(`${rootDir}/source/main.bs`, Position.create(9, 32));
                 names = result.map(x => x.label);
                 expect(names).to.includes('sayHello');
-            });
-        });
-
-        describe('enums', () => {
-            it('gets enum statement completions from global enum', () => {
-                program.addOrReplaceFile('source/main.bs', `
-                    sub Main()
-                        direction.down
-                    end sub
-                    enum Direction
-                        up
-                        down
-                    end enum
-                `);
-                //      |direction.down
-                expectCompletionsIncludes(program.getCompletions('source/main.bs', util.createPosition(2, 25)), [{
-                    label: 'Direction',
-                    kind: CompletionItemKind.Enum
-                }]);
-                //      dire|ction.down
-                expectCompletionsIncludes(program.getCompletions('source/main.bs', util.createPosition(2, 28)), [{
-                    label: 'Direction',
-                    kind: CompletionItemKind.Enum
-                }]);
-                //      direction|.down
-                expectCompletionsIncludes(program.getCompletions('source/main.bs', util.createPosition(2, 33)), [{
-                    label: 'Direction',
-                    kind: CompletionItemKind.Enum
-                }]);
-            });
-
-            it('gets enum member completions from global enum', () => {
-                program.addOrReplaceFile('source/main.bs', `
-                    sub Main()
-                        direction.down
-                    end sub
-                    enum Direction
-                        up
-                        down
-                    end enum
-                `);
-                //      direction.|down
-                expectCompletionsIncludes(program.getCompletions('source/main.bs', util.createPosition(2, 34)), [{
-                    label: 'up',
-                    kind: CompletionItemKind.EnumMember
-                }, {
-                    label: 'down',
-                    kind: CompletionItemKind.EnumMember
-                }]);
-                //      direction.do|wn
-                expectCompletionsIncludes(program.getCompletions('source/main.bs', util.createPosition(2, 36)), [{
-                    label: 'up',
-                    kind: CompletionItemKind.EnumMember
-                }, {
-                    label: 'down',
-                    kind: CompletionItemKind.EnumMember
-                }]);
-                //      direction.down|
-                expectCompletionsIncludes(program.getCompletions('source/main.bs', util.createPosition(2, 38)), [{
-                    label: 'up',
-                    kind: CompletionItemKind.EnumMember
-                }, {
-                    label: 'down',
-                    kind: CompletionItemKind.EnumMember
-                }]);
-            });
-
-            it('gets enum statement completions from namespaced enum', () => {
-                program.addOrReplaceFile('source/main.bs', `
-                    sub Main()
-                        enums.direction.down
-                    end sub
-                    namespace enums
-                        enum Direction
-                            up
-                            down
-                        end enum
-                    end namespace
-                `);
-                //      enums.|direction.down
-                expectCompletionsIncludes(program.getCompletions('source/main.bs', util.createPosition(2, 30)), [{
-                    label: 'Direction',
-                    kind: CompletionItemKind.Enum
-                }]);
-                //      enums.dire|ction.down
-                expectCompletionsIncludes(program.getCompletions('source/main.bs', util.createPosition(2, 34)), [{
-                    label: 'Direction',
-                    kind: CompletionItemKind.Enum
-                }]);
-                //      enums.direction|.down
-                expectCompletionsIncludes(program.getCompletions('source/main.bs', util.createPosition(2, 39)), [{
-                    label: 'Direction',
-                    kind: CompletionItemKind.Enum
-                }]);
-            });
-
-            it('gets enum member completions from namespaced enum', () => {
-                program.addOrReplaceFile('source/main.bs', `
-                    sub Main()
-                        enums.direction.down
-                    end sub
-                    namespace enums
-                        enum Direction
-                            up
-                            down
-                        end enum
-                    end namespace
-                `);
-                //      enums.direction.|down
-                expectCompletionsIncludes(program.getCompletions('source/main.bs', util.createPosition(2, 40)), [{
-                    label: 'up',
-                    kind: CompletionItemKind.EnumMember
-                }, {
-                    label: 'down',
-                    kind: CompletionItemKind.EnumMember
-                }]);
-                //      enums.direction.do|wn
-                expectCompletionsIncludes(program.getCompletions('source/main.bs', util.createPosition(2, 42)), [{
-                    label: 'up',
-                    kind: CompletionItemKind.EnumMember
-                }, {
-                    label: 'down',
-                    kind: CompletionItemKind.EnumMember
-                }]);
-                //      enums.direction.down|
-                expectCompletionsIncludes(program.getCompletions('source/main.bs', util.createPosition(2, 44)), [{
-                    label: 'up',
-                    kind: CompletionItemKind.EnumMember
-                }, {
-                    label: 'down',
-                    kind: CompletionItemKind.EnumMember
-                }]);
-            });
-
-            it('excludes enum member completions from namespace enum', () => {
-                program.addOrReplaceFile('source/main.bs', `
-                    sub Main()
-                        direction.ba
-                    end sub
-                    namespace enums
-                        enum Direction
-                            up
-                            down
-                        end enum
-                    end namespace
-                `);
-                //should NOT find these enums
-                expectCompletionsExcludes(program.getCompletions('source/main.brs', util.createPosition(2, 28)), [{
-                    label: 'up',
-                    kind: CompletionItemKind.EnumMember
-                }, {
-                    label: 'down',
-                    kind: CompletionItemKind.EnumMember
-                }]);
-            });
-
-            it('infers namespace for enum statement completions', () => {
-                program.addOrReplaceFile('source/main.bs', `
-                    namespace enums
-                        sub Main()
-                            direction.down
-                        end sub
-                        enum Direction
-                            up
-                            down
-                        end enum
-                    end namespace
-                `);
-                //          dire|ction.down
-                expectCompletionsIncludes(program.getCompletions('source/main.bs', util.createPosition(3, 33)), [{
-                    label: 'Direction',
-                    kind: CompletionItemKind.Enum
-                }]);
-            });
-
-            it('infers namespace for enum member completions', () => {
-                program.addOrReplaceFile('source/main.bs', `
-                    namespace enums
-                        sub Main()
-                            direction.down
-                        end sub
-                        enum Direction
-                            up
-                            down
-                        end enum
-                    end namespace
-                `);
-                //          direction.do|wn
-                expectCompletionsIncludes(program.getCompletions('source/main.bs', util.createPosition(3, 40)), [{
-                    label: 'up',
-                    kind: CompletionItemKind.EnumMember
-                }, {
-                    label: 'down',
-                    kind: CompletionItemKind.EnumMember
-                }]);
-            });
-
-            it('supports explicit namespace for enum statement completions', () => {
-                program.addOrReplaceFile('source/main.bs', `
-                    namespace enums
-                        sub Main()
-                            enums.direction.down
-                        end sub
-                        enum Direction
-                            up
-                            down
-                        end enum
-                    end namespace
-                `);
-                //          enums.dire|ction.down
-                expectCompletionsIncludes(program.getCompletions('source/main.bs', util.createPosition(3, 38)), [{
-                    label: 'Direction',
-                    kind: CompletionItemKind.Enum
-                }]);
-            });
-
-            it('supports explicit namespace for enum statement completions', () => {
-                program.addOrReplaceFile('source/main.bs', `
-                    namespace logger
-                        sub log()
-                            enums.direction.down
-                        end sub
-                    end namespace
-                    namespace enums
-                        enum Direction
-                            up
-                            down
-                        end enum
-                    end namespace
-                `);
-                //          enums.dire|ction.down
-                expectCompletionsIncludes(program.getCompletions('source/main.bs', util.createPosition(3, 38)), [{
-                    label: 'Direction',
-                    kind: CompletionItemKind.Enum
-                }]);
             });
         });
